Introduce RequestListMappedItem - a simple, compact, consistent object
This refactor has been brewing for a while and was forced when attempting
to improve the item display. Suddenly having to deal with the bloated,
unwieldy, hard-to-navigate PUI record objects in erb templates was more
than I could take.

So, now the plugin provides a first pass mapping of an incoming record
into a RequestItemMappedItem and provides a means for a profile to extend
or override that mapping. By the time you get to setting up form fields
or rendering the item for display, the PUI record is gone and you're
left with the nice little mapped item.

See how it flies anyway.